Add setup_tmate script
The script is setting the tmate server using podman containers. Also
clone the tmate github repo and generate keys for the server if
needed, with tmate configuration file that could be used on host
where tmate session will be running. The script requires to do
some actions under root permissions: set firewalld, .. so you could
be asked to input the password when sudo commands are called.