bashvirt: prevent snapshot changes when external snaps are detected
Added detection of the external snapshots and prevent use of snapshot
functions in such a case to save user from invalid changes.

External snapshots have to be used e.g. in case of VM with UEFI.
But external snaphots are not managed by virsh so simply as internal
snapshots. Probably everyone will find own best way how to deal with
them (google it). One of the newest blogs about is:

    https://fabianlee.org/2021/01/10/kvm-creating-and-reverting-libvirt-external-snapshots/

It's more or less what I am doing in my setup.