bashvirt_ext: Add bash library for work with qemu external snapshots
The original library is not able to handle work with external snapshots
which are required e.g. in case of work with multiple disk devices
or UEFI. This is first attempt to provide at least basic management
level for work with external snapshots. It's far away from ideal state
but it's already usable for very simple cases. Currently the performance
is lowered because of 2times more snapshots than really needed, but
in future it can be definitely improved. It's ok for me right now.